@@ -1,9 +1,8 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import App from './App';
+import { render, screen } from "@testing-library/react";
+import App from "./App";
 
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/learn react/i);
+test("renders learn react link", () => {
+  render(<App />);
+  const linkElement = screen.getByPlaceholderText(/Type/i);
   expect(linkElement).toBeInTheDocument();
 });
